Keep module runtime tests under file-size gate
