+++ .pm/tasks/task_b5b4afa8686b46d58856252de9cea35e.execution.md
@@ -222,3 +222,28 @@ Example:
 - Expected Result: Verification passes, current task is closed, and no current-task closeout data is missing.
 - Actual Result: Verification passed and task yaml now records `status: done`, `last_claim_type: task_complete`, `last_verification_status: verified`, and `last_closed_at: 2026-06-27T14:40:25+08:00`; command exited 1 only because repo-wide pm lint reported unrelated historical `.pm` execution-log debt.
 - Blocker / Next Action: Run `./scripts/pm/workflow-lint.sh --task-uid task_b5b4afa8686b46d58856252de9cea35e --phase pr-ready` and then retry PR preparation.
+
+## 2026-06-27 14:48:30 CST / tpm
+- 完成内容: Debugged and fixed local required gate failure.
+- 遗留事项: Need rerun targeted required command after committing this fix.
+- Action: Reproduced `./scripts/ci-tests.sh required` failure in the recommended targeted required command; all earlier required checks and `oasis7_newapi_bridge_service` tests passed, then `./scripts/check-rust-file-size.sh` failed because `crates/oasis7/src/runtime/world/module_runtime.rs` had grown to 1244 lines.
+- Action: Root cause was this task's new internal tests pushing an implementation file above the 1200-line limit. Moved the `module_runtime` unit tests into `crates/oasis7/src/runtime/world/module_runtime_tests.rs` and left only a `#[cfg(test)] #[path = "module_runtime_tests.rs"] mod tests;` declaration in the implementation file.
+- Validation Command: `wc -l crates/oasis7/src/runtime/world/module_runtime.rs crates/oasis7/src/runtime/world/module_runtime_tests.rs`
+- Expected Result: Both Rust files are below the 1200-line limit.
+- Actual Result: `module_runtime.rs` is 1017 lines; `module_runtime_tests.rs` is 224 lines.
+- Validation Command: `./scripts/check-rust-file-size.sh`
+- Expected Result: Rust file-size gate passes.
+- Actual Result: Passed; oversized code files=0, test files=0.
+- Validation Command: `./scripts/cargo-dev.sh fmt --check`
+- Expected Result: Formatting remains clean after test relocation.
+- Actual Result: Passed.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7 route_tick_to_modules_keeps_schedule_when_due_record_is_missing --lib`
+- Expected Result: Focused P1 regression test still passes after moving test module.
+- Actual Result: Passed; 1 test passed.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7 module --lib`
+- Expected Result: Module regression surface remains intact after moving tests.
+- Actual Result: Passed; 163 tests passed, 1 ignored.
+- Validation Command: `git diff --check`
+- Expected Result: No whitespace errors.
+- Actual Result: Passed.
+- Blocker / Next Action: Commit file-size fix and rerun recommended targeted required validation.
